refactor(helpers): refactor helpers module - Refactors helper module into separate files in "helpers" folder - Existing "helpers" file left as facilitator to split helper module - Report "generate_raw_exceptions" function split into helper file
